+++ docs/changelog.rst
@@ -58,6 +58,10 @@ Bug fix release.
   one. Particularly relevant for Mopidy-Scrobbler users, as before it was
   essentially unusable. (Fixes: :issue:`1456`, PR: :issue:`1534`)
 
+- Models: Fix encoding error if :class:`~mopidy.models.fields.Identifier`
+  fields, like the ``musicbrainz_id`` model fields, contained non-ASCII Unicode
+  data. (Fixes: :issue:`1508`, PR: :issue:`1546`)
+
 - File: Ensure path comparison is done between bytestrings only. Fixes crash
   where a :confval:`file/media_dirs` path contained non-ASCII characters.
   (Fixes: :issue:`1345`, PR: :issue:`1493`)
@@ -409,7 +413,7 @@ Bug fix release.
   proceed startup. (Fixes: :issue:`1248`, PR: :issue:`1254`)
 
 - Stream: Fix bug in new playlist parser. A non-ASCII char in an urilist
-  comment would cause a crash while parsing due to comparision of a non-ASCII
+  comment would cause a crash while parsing due to comparison of a non-ASCII
   bytestring with a Unicode string. (Fixes: :issue:`1265`)
 
 - File: Adjust log levels when failing to expand ``$XDG_MUSIC_DIR`` into a real